feat: systemd unit for apx-vso-pico
Adds a systemd unit for managing apx-vso-pico, which automatically
starts the container on startup and shuts it down on poweroff. When
enabled by the user (new installations should do by default with
first-setup), this service avoids having to wait for the container
to start when opening a terminal for the first time, as well as
drastically reduces the poweroff time.